// apps/admin/src/lib/api/generated/models/productInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * DevFeed Admin API
 * Private administration API. OIDC sessions and CSRF protection required.
 * OpenAPI spec version: 0.0.50
 */
import type { Evidence } from './evidence';
import type { ProductInputPricing } from './productInputPricing';

export interface ProductInput {
  /** @maxLength 300 */
  attribution?: string;
  /**
     * @minLength 10
     * @maxLength 5000
     */
  description: string;
  /** @maxItems 10 */
  evidence?: Evidence[];
  /**
     * @minLength 1
     * @maxLength 200
     */
  external_id: string;
  listing_url: string;
  /**
     * @minLength 1
     * @maxLength 200
     */
  name: string;
  pricing?: ProductInputPricing;
  product_url: string;
  /**
     * @minLength 1
     * @maxLength 200
     */
  provider: string;
  /**
     * @maxItems 20
     * @items.minLength 1
     * @items.maxLength 200
     */
  technologies?: string[];
}