// apps/admin/src/lib/api/generated/models/nickProduct.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * DevFeed Admin API
 * Private administration API. OIDC sessions and CSRF protection required.
 * OpenAPI spec version: 0.0.50
 */

export interface NickProduct {
  /** @maxLength 5000 */
  description?: string;
  /**
     * @minLength 1
     * @maxLength 200
     */
  name: string;
  /** @maxLength 100 */
  pricing?: string;
  productUrl: string;
  /**
     * @minLength 1
     * @maxLength 200
     */
  slug: string;
  /** @maxLength 5000 */
  tagline?: string;
  url: string;
}